requirements.txt
requests     
openai       
gitpython    
python-dotenv